+++ iotdb-client/cli/src/test/java/org/apache/iotdb/cli/AbstractCliTest.java
@@ -41,8 +41,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.when;
 
-public class AbstractCliIT {
-  private static Logger logger = LoggerFactory.getLogger(AbstractCliIT.class);
+public class AbstractCliTest {
+  private static Logger logger = LoggerFactory.getLogger(AbstractCliTest.class);
   @Mock private IoTDBConnection connection;
 
   @Mock private IoTDBDatabaseMetadata databaseMetadata;
